requirements.txt
# This must be installed before pygrib, since pygrib's setup.py is broken
numpy==1.7.1


statsd-client==1.0.2
gevent==0.13.8
pygrib==1.9.6
pyproj==1.9.3

